Only create root directory on installation.
Update uninstall logic so that it detects that `~/.uwe` does not
exist and prints a message saying the tools are not installed and how to
install them.

Closes #414.

+++ components/release/src/uninstall.rs
@@ -9,15 +9,26 @@ use crate::Result;
 pub async fn uninstall() -> Result<()> {
     let dir = dirs::root_dir()?;
 
-    print!("Uninstall {}? (y/n) ", dir.display());
+    if !dir.exists() || !dir.is_dir() {
+        info!("Not installed {}", dir.display());
+        info!("");
+        info!("To install the platform tools run:");
+        info!("");
+        info!("curl https://releases.uwe.app/install.sh | sh");
+        info!("");
+        return Ok(())
+    }
+
+    print!(" Uninstall {}? (y/n) ", dir.display());
     io::stdout().flush()?;
 
     let stdin = io::stdin();
     for line in stdin.lock().lines() {
         let response = line.unwrap();
         if response == "y" || response == "yes" {
+            utils::terminal::clear_previous_line()?;
             fs::remove_dir_all(&dir)?;
-            info!("Uninstalled {}", dir.display());
+            info!("Uninstalled {} ✓", dir.display());
         }
         break;
     }
